Addressing Dave's reply on name table rules for NAMEID_FONT_FAMILY_NAME:
For PID=1 then it will always be the same across all styles.

For PID=3 then it will only be the same for the 4 'core' or 'ribbi' styles (Regular, Italic, Bold, Bold Italic) and for the others it will be Family + Weight, eg "Italic" must never appear here.

Example: "Family Sans Thin" or "Family Medium"