Implementing a class to abstract the 2 kinds of targets:
(A) fullpath strings for local filesystem access of font data
or (B) in-memory file-like TTFont objects with accompaining metadata dictionary (as the info is provided by our webapp)

The class works as a façade hiding the details from (and thus simplifying) the main FontBakery TTF checker script.

This is work in progress related to issue #1181.