check-ttf.py: Fixed IOError for function check_with_msfontvalidator.
If the user doesn't have MS Font Validator installed, the file report.xml
will not be generated. The variable xml_report will not be able to load this
file. By adding this variable and subsequent statements to the try and accept
block, it will run the exception and not crash the script, if the user does
not have MS Validator installed.