using some shell-foo I extended the list with all gfonts
font-family names, based on the foldernames in the google/fonts repo.
We'll still have to fill the gaps in this table...

Number of font repositories in this file:
"TTX": 18
"OK": 46
"UFO": 17
"GH-PAGES": 2
"NOTE": 13
"GFONTS-REPO": 1
"OTF": 15
"404-ERROR": 1
"?": 802
"SOURCE-ONLY": 1
From the total of 916 repos, 59 are enabled.
Enabled means: "OK" or "NOTE" status.

(issue #1241)