Making check implementations more self-contained
There's no point anymore in keeping constants at the fontbakery.constants module unless they're used by more than a single check.
Also, I've migrated most of the import statements into the check implementation bodies so that copying and pasting them into external python scripts would be easier in case people may be interested in manually reusing some of the routines.

The previous way we did things with constants in a separate module was useful to reduce code complexity and improve legibility. But now that we have a pretty modular design for the checks it makes much more sense to do things differently. Also, the "all imports on the top" practice comes from when we had the requirement to keep fontbakery as a standalone single-script.

There are still some constants being loaded from the external module. I'll first finish porting the ~150 checks, so that in the end I can more easily see which constants are used only once and declare them inline, while only keeping on the separate module the ones that are reused in multiple checks.

(issue #1388)