Clarify wording on test/155 to make it clear that
the copyright notice entries on name table are the reference values which are expected to be found on METADATA.pb font.copyright field (not the other way around).

This does not alter the test implementation. It only changes the wording of the log messages and docstrings and comments/messages on the corresponding self-test.
(issue #183)