bin/fontbakery-check-gf-github.py: Improved performance.
In the previous implementation, if specified date ranges were longer than a month, the results
would take minutes to generate. In the new implementation, this is reduced to under a minute at
most.

This pr is quite rough because @davelab6 would prefer these analytics to form part of the fb dashboard
we can work on a better implementation then