Fix check 174 on Windows by not using the filesystem
Having TTFont save() to the named temporary file throws an IOError on Windows.

The docs for tempfile.NamedTemporaryFile say: "Whether the name can be used to
open the file a second time, while the named temporary file is still open,
varies across platforms (it can be so used on Unix; it cannot on Windows NT or
later)."

Solved by using a TemporaryFile, since TTFont.save() also accepts a writable
file object.