specify style as a condition to check/020 (usWeightClass)
This fixes a crash when running it on a variable font.
Alternatively one could use conditions=['not is_variable_font']
but I think that having 'style' not only dependency-injected but also
explicitely listed as a condition is the more general and safe way
to go here. After all, the code will crash with style=None, so it is
indeed strictly required.

(issue #1964)