general_test.py: test ots with invalid file
It would have been better to test with an actual font, I know.
But I was lazy ;)
We also need to add a test for the WARN case, i.e. when the ots-sanitize
passes (returns 0) but something is printed to stderr (signalling that
some warning message was emitted).
This happens e.g. when one of the non-mandatory tables fails to sanitize,
but in that case they are simply dropped, and the font as a whole still
sanitizes.