Bump up requests version to mitigate CVE
Even though I think our usage of requests in Font Bakery likely does not expose our users to this specific vulnerability, I think it is harmless to require a newer version of the requests module, which is the suggested mitigation for the issue.

CVE-2018-18074
The Requests package through 2.19.1 before 2018-09-14 for Python sends an HTTP Authorization header to an http URI upon receiving a same-hostname https-to-http redirect, which makes it easier for remote attackers to discover credentials by sniffing the network.

(issue #2135)