Deprecating FontForge checks
Python 3 FontForge support is messy and likely misconfigured for all FontBakery users.
It is non-trivial to setup a system in which one can properly run the fontforge checks with FontBakery.
So we'll have to put some significant effort upstream (which probably includes building and distributing python3 enabled binaries) if we ever try to bring back the fontforge checks to FontBakery.