Update seems_monospaced conditions
When a font doesn’t have ASCII characters like A-Za-z, but at least
80% of ASCII character glyphs with the same width, for example if it has
tabular figures, fontbakery would assume it is a monospace font and
com.google.fonts/check/monospace would fail.

This extends the conditions so when less than 80% ASCII characters are in the
font, the glyphs of characters should have one of two widths, excluding glyphs
of control characters, mark glyphs or zero-width glyphs.