check/varfont/grade_reflow - Cleanup output
Use a python set (instead of a python list) in order to eliminate multiple reporting of the same glyphs.
(issue #3561)

Also fix reference of the fix from the previous commit to the correct issue (#3566)
on the CHANGELOG file. (unfortunately the bad reference will remain incorrect in the git log message)