fix parsing at production_metadata condition
The json data available at http://fonts.google.com/metadata/fonts changed in a way that breaks the parsing at the production_metadata condition.

In the past there were 5 spurious characters in the beginning of that content and we had to work-around that. Such a hack is not needed anymore, and it is currently causing check ERRORs.
(issue #3661)