Override freetype_rasterizer to make it mandatory for Google Fonts
At 8bf12c8, I noticed that it would be wise for the Google Fonts profile to also override the freetype_rasterizer check from SKIP (if freetype dependency is not installed) to a FAIL, just like Adobe Fonts does.

We will also need to setup the google/fonts CI rules so that we install freetype before running fontbakery on font PRs.

The check, on the Universal Profile, emits a SKIP by default because it is considered an optional check. But I think it should be mandatory for Google Fonts.
(issue #3871)