Deprecate com.google.fonts/check/listed_on_gfonts
Removed check from Google Fonts profile.
It did not pass for any new families and was not deemed to be a useful check by the onboarding team.
The WARN was actually considered annoying.

The listed_on_gfonts_api condition remains, though, as it is still is use in other checks.
(issue #3220)