New check: glyphsets/shape_languages
This check uses a heuristic to determine which GF glyphsets a font supports.
Then it checks the font for correct shaping behaviour for all languages in those glyphsets.
FAILs if no glyphset is found to be supported >80%.

com.google.fonts/check/glyphsets/shape_languages (experimental)
Added to the Google Fonts Profile
(issue #4147)