New Check: metadata/primary_script
This infers the font's primary script and then
checks that's properly set in METADATA.pb.

This is an educated guess based on the number of glyphs per
script in the font and should be taken with caution.

Added to the Google Fonts Profile
com.google.fonts/check/metadata/primary_script
(issue #4109)