New check: Arabic letter high hamza isn't classified as a mark.
Many fonts incorrectly treat ARABIC LETTER HIGH HAMZA (U+0675) as a variant of
ARABIC HAMZA ABOVE (U+0654) and make it a combining mark of the same size.

But U+0675 is a base letter and should be a variant of ARABIC LETTER HAMZA
(U+0621) but raised slightly above baseline.

Not doing so effectively makes the font useless for Jawi and
possibly Kazakh as well.

Added to the Universal Profile
com.google.fonts/check/arabic_high_hamza (experimental)
(issue #4290)