Fix shaping regression HTML output
The code is writing out HTML with <style> tag, but the markdown to HTML conversion escapes the <style> tag breaking the rendering. Simplify the whole thing by using markdown syntax instead and adding inline styling to the SVG. The output result is slightly different than before it broke, but I think the difference is OK.

Also, don’t show test header twice in shaping tests. It is shown next to the FAIL line, then again right after it. Drop the second one and simplify the code.

(PR #4406)