[CODE-REVIEW] New check: tabular_kerning
Check that tabular numerals and symbols have no kerning.

Added to the Universal Profile
EXPERIMENTAL - com.google.fonts/check/tabular_kerning

(issue #4440)
----

The vharfbuzz and uharfbuzz dependencies are no longer profile-specific, as this new
check that uses them belongs to the Universal profile, so the deps are
now included by default on a vanilla fontbakery installation.