Add HTML templating
Currently we mix a lot of HTML and code in the same reporter, leading to
messy code that's hard to improve. This uses Jinja2 to feed the
JSON output from the serialize reporter into a bunch of HTML templates.

The reporter code is a lot simpler, and as a bonus, we make it easier
for profile owners to change the look-and-feel of their reports.

(PR #4460)