Split "googlefonts" profile into manageable files
Currently the check definitions for the googlefonts profile are in a 7 kloc monster module. There are operating systems smaller than that. Having all those checks in a single file makes it hard to see how things relate to one another, whether there is duplication, where new checks should be placed, and so on.

This groups the definitions into functional modules. It also begins the process of splitting conditions, cached values and utility functions.

(PR #4501)