Make hinting_stats have its own ttFont instance
Fixes issue #4525. The problem is that conditions are meant to be invariant. In theory, the result of conditions were cached all along; but actually, they weren't! After #4517 conditions really are cached, so anything which modifies font.ttFont also changes it for future tests. We need to make sure they're not doing that.