Use dir() instead of getattr()
We currently use hasattr to determine what conditions are available on a testable (file or whole collection), to see whether or not to route a testable to a check. For a cached property, hasattr actually gets the value of the property! dir() does not, so we use that instead, although it requires a little bit of tweaking to the mock object.

I'm not sure this actually makes fontbakery faster, but it certainly makes it feel faster, because we do the network conditions etc. when they are used, not at startup...

Note: Looking at the CI runs it's about 10% faster.

(PR #4558)