Correctly generated the list of ligature pairs in the font
If the font has a ligature table with more than one entry, e.g.:

    {"f": [["i"], ["l"]]}

The code would generate the kern pairs:

    [("f", "i"), ("i", "l")]

This because the loop was changing `first` from `f` to `i` while
iterating on the ligatures. This was meant to handle cases like:

    {"f": [["f", "i"], ["f", "l"]]}

to generate:

    [("f", "f"), ("f", "i"), ("f", "l")]

Both are now correctly handled.

Fixes https://github.com/fonttools/fontbakery/issues/2227