Eager-load fonts while protected by a lock
I'm eager-loading all tables of the ttFont under test while we're opening the ttFont in a cached_property, because that decorator uses proper locking. In my previously bad runs with --auto-jobs, this now completes the run without errors, however it introduces a long pause at the start of execution before running any check, presumably while the whole font gets loaded.

An alternative solution would be to introduce locking in fontTools, so that lazy-loading a single table has a lock on that entry in the TTFont table cache. That way, only the tables actually needed by tests would get loaded.

It could be interesting to try also with deepcopy; and yet another idea could be that the getter would need to be cached per-thread instead of globally, i.e. use thread-local storage for the caching of cached_property, and so have one ttFont instance per thread. There would be duplicated work to load tables but each thread would have its own copy so hopefully they wouldn't step on each other's toes, and possibly each thread would only load the few tables that the checks running in that thread will use.

But for now we'll simply eager-load the fonts when the user enables multi-threading.

(issue #4638)