Re-enable check/legacy_accents
Re-enabled the check, but dropped complaints about legacy accents as part of glyph compositions.

The two messages `legacy-accents-width` and `legacy-accents-gdef` are critical and should not have been disabled in my opinion.

Since the check was disabled rather than outright deleted, I interpret from this that back then there was no time to look into the matter in detail, and so I took the initiative to completely remove the `legacy-accents-component` WARN part but otherwise re-enabled the check as it was to still yield FAIL on the two other messages.

com.google.fonts/check/legacy_accents
On the Universal profile.

(issue #4567)