Extend "regular coord" conditions
to accept "Regular Italic" style name

- regular_wght_coord
- regular_wdth_coord
- bold_wght_coord

On shared conditions.

While it is common for typefaces to have a Regular Italic style called simply "Italic," it is also common for typefaces to label this style as "Regular Italic" – and arguably, this is more user-friendly in style menus that list many weights, plus Italics.

There are multiple checks that rely on conditions that find the "Regular" instance of an Italic variable font, then use this to determine whether that instance has the correct axis values for weight, width, and optical size. This is great! However, this leads to a lot of false WARNs/FAILs if such a font uses the style name "Regular Italic," instead of simply "Italic."

A bit more context: The shorter name (simply "Italic"), as I understand it, is used to help apps and systems that use style names for RIBBI style linking. However, as I understand it, these systems should continue to work so long as NameID 2 is set to just "Italic" (plus a few factors in tables other than the name table). If a named instance is "Regular Italic," or if NameID 17 is "Regular Italic," it is my understanding and experience that things will still work fine. If I am incorrect in this, I am happy to be corrected.

(PR #4741)