Warn about paths without bounds (instead of crashing)
Bodoni Moda's "OE" glyph looks like this:

    <TTGlyph name="OE" xMin="80" yMin="0" xMax="1850" yMax="1500">
      <contour>
        <pt x="784" y="1461" on="1"/>
      </contour>
      <contour>
        <pt x="311" y="750" on="1"/>
        <pt x="311" y="886" on="0"/>
        <pt x="356" y="1141" on="0"/>
        ...

Yep, that's a single-point contour. This currently crashes the outline direction check.
But a single point contour is probably something we want to warn about anyway, so it
seems a good idea to (a) make it not crash and (b) make it warn about the contour instead.

com.google.fonts/check/outline_direction
On the Outlines profile.

(PR #4745)