Speed up expected_font_names
In expected_font_names we do a deep copy of the font because we're going to modify it
and we don't want to mess up further tests. But for big fonts this is very slow, and
we use expected_font_names a few times. We don't need glyf and gvar and GPOS and
everything else to work out what names the font should have. Instead, let's only copy
the tables that we need to run build_name_table.

(PR #4765)