'aat' check moved to universal profile
'aat' check ('unwanted aat tables') was included in all vendor-specific
profiles, except Microsoft's. So it was moved to the Universal profile
and added to the "pending_review" list of the Microsoft profile.

Also, its rationale text was updated to remove references to Google
Fonts. And the list of unwanted tables was sorted alphabetically.