dismantling checks/universal directory
We want to treat the Lib/fontbakery/checks directory as a general pool of checks, so we don't want to have any resemblance of profile assignments in that directory.

Profile definition should happen only at Lib/fontbakery/profiles

There will be, though, a few vendor-specific files/subdirectories below
Lib/fontbakery/checks, but only when we are sure that the checks are
really specific to a vendor.