Clarify relation between notofonts and adobefonts profiles.
A few checks from notofonts were not marked "PERMANENTLY EXCLUDED"
on Adobe Fonts profile while others were just commented out but
not explicitely marked for permanent exclusion.

At the same time, Adobe Fonts was not using the import_profile feature
to include checks from the notofonts one.

So, this seems to indicate that what actually happened was that Adobe
inspected the full catalog of checks accross all of the FontBakery
codebase and "cherry-picked" the ones they wanted, which is totally
fine, but resulted in these comments about exclusion of checks that
had not even been included at first.

So here I am deleting all lines with these permanently excluded
checks. And, naturally, not including them on the exclude_checks or
pending_review lists.

The only check from notofonts that was picked by Adobe was
`cmap/format_12`, so I will (in a future commit) move it to Universal
and then add it to the pending_review list of other vendor profiles.

And these other two checks are being placed on adobe's pending_review
list because they had not been explicitly excluded, so maybe they are
still considered useful here:

- cmap/alien_codepoints
- unicode_range_bits

----
I hope that, by being a bit verbose here, I made it clear to others (and even
to myself in the future) what was the reasoning behing this change.