Improve description of TypeNetwork profile:
Similarly to the changes made on the Adobe Profile at commit 2ca4107757bc6f0eb3158e353fde365a1730856a

Here I'm deleting all PERMANENTLY EXCLUDED entries. All new checks
implemented in FontBakery will be added to the pending_review list of
the vendor profiles, so that it is not needed anymore to keep track with
these comments.

Also, I'm removing comments like "OVERRIDDEN to WARN" which are
redundant with the actual override declarations present in the file. It
was also error prone, as there was the risk of having changes made to
override declarations while forgetting to keep these comments in sync.