Use / instead of : on check-IDs
I had originally considered using : to separate the optional vendor
prefixes (such as "typenetwork:", "fontbureau:", etc.) but now I think
it is better to simply use a forward slash, because that colon might
convey an idea of some special/magic handling of the checks that share
identical names after the colon. But that's not the case. These
vendor-specific check IDs do not currently have any automatic feature
gathering them.

Using forward slashes makes it clear these are simple names for the
checks.