opentype/italic_angle was affecting empty_letters
Fix the problem by working on a copy of the ttFont object.

Here I removed the empty_letters check from Type Network's
pending_review list because it was marked as "the check is
broken". So, I assume that the check was considered useful
and would be included if it was not misbehaving.