Fix mock glyph removal now that check no longer mutates TTFont
The previous approach relied on the `glyphOrder` property being
populated as a side-effect of the "unreachable-glyphs" check, but this
no longer happens now that the accidental mutation in the check has been
removed.

This commit avoids using the internal property entirely by using the
`setGlyphOrder()` function, and ensuring that the font is fully loaded
prior to this to avoid the glyph being referenced during a later load
after it has already been removed from the order.

If this causes issues again, it may be more resilient for us to use
fontTools' subsetter library to implement this pytest.

(PR #4835)